made smtrat master branch compilable with clang and mac os

+++ src/solver/parser/Driver.cc
@@ -878,7 +878,7 @@ namespace smtrat
      */
     Rational Driver::getRational( string* _numString ) const
     {
-        unsigned pos = _numString->find('.');
+        size_t pos = _numString->find('.');
         if( pos != string::npos )
         {
             unsigned numDecDigits = _numString->size()-pos-1;
